Prevent a logged-off user's local observations from being edited by the current user (#69)
* WIP - only show local observations for the current user

this ensures that only the current user can see and even try to update the local observations. Without this, a device can have multiple accounts, with only one being logged into, and yet show observations for accounts that aren't logged into. This causes permission denials when an observation is edited for an account that isn't logged in - which makes sense

* WIP - updated to version 3.0.5